src/processor-lambda/requirements.txt
boto3
botocore
azure-storage-blob
